@@ -1,24 +1,29 @@
 
 buildscript {
     repositories {
+        google()
+        maven { url "https://maven.google.com" }
         jcenter()
-        maven { url 'https://jitpack.io' }
     }
 
     dependencies {
-        classpath 'com.android.tools.build:gradle:2.2.3'
+        classpath 'com.android.tools.build:gradle:3.2.1'
     }
 }
 
 apply plugin: 'com.android.library'
 
+def safeExtGet(prop, fallback) {
+    rootProject.ext.has(prop) ? rootProject.ext.get(prop) : fallback
+}
+
 android {
-    compileSdkVersion 25
-    buildToolsVersion '25.0.3'
+    compileSdkVersion safeExtGet("compileSdkVersion", 27)
+    buildToolsVersion safeExtGet("buildToolsVersion", "28.0.3")
 
     defaultConfig {
-        minSdkVersion 16
-        targetSdkVersion 25
+        minSdkVersion safeExtGet("minSdkVersion", 16)
+        targetSdkVersion safeExtGet("targetSdkVersion", 27)
         versionCode 1
         versionName "1.0"
     }
@@ -29,14 +34,12 @@ android {
 
 repositories {
     mavenCentral()
-    maven { url 'https://jitpack.io' }
+    maven { url "https://maven.google.com" }
 }
 
-configurations.all {
-    resolutionStrategy.cacheChangingModulesFor 0, 'seconds'
-}
 
 dependencies {
-    compile 'com.facebook.react:react-native:+'
-    compile 'com.github.congnguyen91:TapTargetView:30b49da'
+    implementation 'com.facebook.react:react-native:+'
+    implementation 'com.github.LuisBonsembiante:TapTargetView:6ebfd66'
+    // implementation 'com.getkeepsafe.taptargetview:taptargetview:1.12.0'
 }
