@@ -1,15 +1,19 @@
 package ui.apptour;
 
+import android.annotation.TargetApi;
 import android.app.Activity;
 import android.content.pm.PackageManager;
 import android.content.res.Resources;
 import android.graphics.Color;
 import android.graphics.Rect;
+import android.graphics.Typeface;
 import android.util.Log;
-import android.support.annotation.Nullable;
+import androidx.annotation.Nullable;
 import android.app.Dialog;
 import android.view.View;
 import android.app.AlertDialog;
+import android.content.Context;
+import android.content.res.AssetManager;
 
 import com.facebook.react.uimanager.UIManagerModule;
 import com.facebook.react.uimanager.UIBlock;
@@ -38,13 +42,17 @@ import java.util.ArrayList;
 import java.util.List;
 
 
+
 public class RNAppTourModule extends ReactContextBaseJavaModule {
 
     private final ReactApplicationContext reactContext;
+    private AssetManager AssetManager;
 
     public RNAppTourModule(ReactApplicationContext reactContext) {
         super(reactContext);
+
         this.reactContext = reactContext;
+        this.AssetManager = reactContext.getApplicationContext().getAssets();
     }
 
     @Override
@@ -178,6 +186,7 @@ public class RNAppTourModule extends ReactContextBaseJavaModule {
         String skipButtonBackgroundColor = null;
         String textColor = null;
         String dimColor = null;
+        String fontFamily = null;
 
         if (props.hasKey("description") && !props.isNull("description")) {
             description = props.getString("description");
@@ -208,6 +217,9 @@ public class RNAppTourModule extends ReactContextBaseJavaModule {
         }
         if (props.hasKey("dimColor") && !props.isNull("dimColor")) {
             dimColor = props.getString("dimColor");
+        }
+        if (props.hasKey("fontFamily") && !props.isNull("fontFamily")) {
+            fontFamily = props.getString("fontFamily");
         }
 
 
@@ -340,6 +352,10 @@ public class RNAppTourModule extends ReactContextBaseJavaModule {
             targetView.textColorInt(Color.parseColor(textColor));
         if (dimColor != null && dimColor.length() > 0)
             targetView.dimColorInt(Color.parseColor(dimColor));
+        if (fontFamily != null && fontFamily.length() > 0) {
+            Typeface font = Typeface.createFromAsset(this.AssetManager, "fonts/" + fontFamily + ".ttf");
+            targetView.textTypeface(font);
+        }
 
         targetView.outerCircleAlpha(finalOuterCircleAlpha);
         targetView.titleTextSize(finalTitleTextSize);
