@@ -4,83 +4,75 @@
  * @flow
  */
 
-import React, { Component } from 'react';
+import React, { Component } from 'react'
 import {
   Platform,
   StyleSheet,
   Text,
   View,
   Button,
-  DeviceEventEmitter,
-} from 'react-native';
+  DeviceEventEmitter
+} from 'react-native'
 
-import { AppTour, AppTourSequence, AppTourView } from 'react-native-app-tour';
+import { AppTour, AppTourSequence, AppTourView } from 'react-native-app-tour'
 
-import Top from './components/Top';
-import Center from './components/Center';
-import Bottom from './components/Bottom';
+import Top from './components/Top'
+import Center from './components/Center'
+import Bottom from './components/Bottom'
 
 const instructions = Platform.select({
   ios: 'Press Cmd+R to reload,\n' + 'Cmd+D or shake for dev menu',
   android:
     'Double tap R on your keyboard to reload,\n' +
-    'Shake or press menu button for dev menu',
-});
+    'Shake or press menu button for dev menu'
+})
 
 export default class App extends Component<{}> {
   constructor(props) {
-    super(props);
+    super(props)
 
-    this.appTourTargets = [];
+    this.appTourTargets = []
   }
 
   componentWillMount() {
-    this.unregisterAppTourEvent();
-    this.registerFinishAppTour();
+    this.registerSequenceStepEvent()
+    this.registerFinishSequenceEvent()
   }
 
   componentDidMount() {
     setTimeout(() => {
-      let appTourSequence = new AppTourSequence();
+      let appTourSequence = new AppTourSequence()
       this.appTourTargets.forEach(appTourTarget => {
-        appTourSequence.add(appTourTarget);
-      });
-      AppTour.ShowSequence(appTourSequence);
-    }, 1000);
+        appTourSequence.add(appTourTarget)
+      })
+
+      AppTour.ShowSequence(appTourSequence)
+    }, 1000)
   }
 
-  componentWillUnmount() {
-    this.unregisterAppTourEvent();
-  }
-
-  registerFinishAppTour = () => {
-    this.unregisterAppTourEvent();
-    this.finishAppTour = DeviceEventEmitter.addListener('onFinishSequenceEvent', (e: Event) => {
-      console.log(e);
-    });
-    this.cancelAppTour = DeviceEventEmitter.addListener('onCancelStepEvent', (e: Event) => {
-      console.log(e);
-    });
-    this.stepAppTour = DeviceEventEmitter.addListener(
+  registerSequenceStepEvent = () => {
+    if (this.sequenceStepListener) {
+      this.sequenceStepListener.remove()
+    }
+    this.sequenceStepListener = DeviceEventEmitter.addListener(
       'onShowSequenceStepEvent',
       (e: Event) => {
-        console.log(e);
+        console.log(e)
       }
-    );
-  };
+    )
+  }
 
-  unregisterAppTourEvent = () => {
-    if (this.finishAppTour) {
-      this.finishAppTour.remove();
+  registerFinishSequenceEvent = () => {
+    if (this.finishSequenceListener) {
+      this.finishSequenceListener.remove()
     }
-    if (this.cancelAppTour) {
-      this.cancelAppTour.remove();
-    }
-    if (this.stepAppTour) {
-      this.stepAppTour.remove();
-    }
-};
-
+    this.finishSequenceListener = DeviceEventEmitter.addListener(
+      'onFinishSequenceEvent',
+      (e: Event) => {
+        console.log(e)
+      }
+    )
+  }
 
   render() {
     return (
@@ -88,23 +80,23 @@ export default class App extends Component<{}> {
         <Top
           style={styles.top}
           addAppTourTarget={appTourTarget => {
-            this.appTourTargets.push(appTourTarget);
+            this.appTourTargets.push(appTourTarget)
           }}
         />
         <Center
           style={styles.center}
           addAppTourTarget={appTourTarget => {
-            this.appTourTargets.push(appTourTarget);
+            this.appTourTargets.push(appTourTarget)
           }}
         />
         <Bottom
           style={styles.bottom}
           addAppTourTarget={appTourTarget => {
-            this.appTourTargets.push(appTourTarget);
+            this.appTourTargets.push(appTourTarget)
           }}
         />
       </View>
-    );
+    )
   }
 }
 
@@ -112,15 +104,15 @@ const styles = StyleSheet.create({
   container: {
     flex: 1,
     flexDirection: 'column',
-    justifyContent: 'space-between',
+    justifyContent: 'space-between'
   },
   top: {
-    flex: 1,
+    flex: 1
   },
   center: {
-    flex: 1,
+    flex: 1
   },
   bottom: {
-    flex: 1,
-  },
-});
+    flex: 1
+  }
+})
