@@ -1,3 +1,7 @@
+/**
+ * @format
+ */
+
 import 'react-native';
 import React from 'react';
 import App from '../App';
@@ -6,7 +10,5 @@ import App from '../App';
 import renderer from 'react-test-renderer';
 
 it('renders correctly', () => {
-  const tree = renderer.create(
-    <App />
-  );
+  renderer.create(<App />);
 });
