@@ -1,4 +1,9 @@
-import { AppRegistry } from 'react-native';
-import App from './App';
+/**
+ * @format
+ */
 
-AppRegistry.registerComponent('Example', () => App);
+import {AppRegistry} from 'react-native';
+import App from './App';
+import {name as appName} from './app.json';
+
+AppRegistry.registerComponent(appName, () => App);
