@@ -5,7 +5,7 @@
 @interface MutableOrderedDictionary<__covariant KeyType, __covariant ObjectType> : NSDictionary<KeyType, ObjectType>
 @end
 
-@interface RNAppTour : NSObject<RCTBridgeModule> {
+@interface RNAppTour : NSObject<RCTBridgeModule, MaterialShowcaseDelegate> {
     MutableOrderedDictionary *targets;
     UIColor *targetOriginalColor;
 }
